+++ src/test/resources/1st-workflow-mod.cwl
@@ -1,0 +1,24 @@
+cwlVersion: v1.0
+class: Workflow
+inputs:
+  inp: File
+  ex: string
+
+outputs:
+  - id: classout
+    type: File
+    outputSource: compile/classfile
+
+steps:
+  untar:
+    run: tar-param.cwl
+    in:
+      tarfile: inp
+      extractfile: ex
+    out: [example_out]
+
+  compile:
+    run: arguments.cwl
+    in:
+      src: untar/example_out
+    out: [classfile]
